+++ lib/core/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod bind;
 pub mod buffer;
+pub mod component;
 pub mod frame;
 pub mod rpc;
 pub mod visual;